CPR-199 move reusable functions into IntegrationTestBase
